Removed a tonne of stuff related to the old webhooks processing
